Debugged and finished simulation function.
The theta parameters were missing values for the G factor, thus leading to
incompatible matrix shapes when trying to do matrix multiplication. Therefor,e
an extra dimension for the G factor was added. Furthermore, the results are now
returned in the form of a list giving back the parameters, the true and estimated
coefficients as well as their deviation.